파일 업로드 완료함

+++ sm_chatting_back/src/main/java/com/sbb/sm_chatting/DTO/Message.java
@@ -7,10 +7,9 @@ import java.time.LocalDateTime;
 @Data
 public class Message {
     private String talkroom_id;
-    private int senduserid;
+    private long senduserid;
     private String content;
     private LocalDateTime regdate;
-
     private String files;
 
     // 생성자
@@ -19,7 +18,7 @@ public class Message {
     }
 
     // 생성자 작성으로 payload가 들어오자마자 Message 셋팅이 가능하다.
-    public Message(String talkroom_id, int senduserid, String content, LocalDateTime regdate, String files) {
+    public Message(String talkroom_id, long senduserid, String content, LocalDateTime regdate, String files) {
         this.talkroom_id = talkroom_id;
         this.senduserid = senduserid;
         this.content = content;
